Har ændret alt fra long til integers da det irriterede mig at der var forskellige datatyper. Har lavet en forside med links til de forskellige ressourcer. Har fikset så man nu kan se en bookinglist i hver user, hvis en booking er tilknyttet brugeren.

+++ src/main/java/com/grp2055/restbackend/controllers/BookingController.java
@@ -27,7 +27,7 @@ public class BookingController {
 
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Booking getBookingById(@PathVariable Long id){
+    public Booking getBookingById(@PathVariable int id){
         return bookingService.findBookingById(id);
     }
 
@@ -39,7 +39,7 @@ public class BookingController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void deleteBooking(@PathVariable Long id){
+    public void deleteBooking(@PathVariable int id){
         bookingService.deleteBooking(id);
     }
