+++ src/main/java/com/grp2055/restbackend/service/BookingServiceImpl.java
@@ -4,6 +4,7 @@ import com.grp2055.restbackend.domain.Booking;
 import com.grp2055.restbackend.repositories.BookingRepo;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -15,7 +16,7 @@ public class BookingServiceImpl implements BookingService {
     }
 
     @Override
-    public Booking findBookingById(Long id) {
+    public Booking findBookingById(int id) {
         return bookingRepo.findById(id).get();
     }
 
@@ -30,7 +31,7 @@ public class BookingServiceImpl implements BookingService {
     }
 
     @Override
-    public void deleteBooking(Long id) {
+    public void deleteBooking(int id) {
         bookingRepo.deleteById(id);
     }
 }
