+++ src/main/java/com/grp2055/restbackend/service/RoomServiceImpl.java
@@ -15,7 +15,7 @@ public class RoomServiceImpl implements RoomService {
     }
 
     @Override
-    public Room findRoomById(Long id) {
+    public Room findRoomById(int id) {
       return roomRepo.findById(id).get();
     }
