+++ src/main/java/com/grp2055/restbackend/service/UserServiceImpl.java
@@ -17,7 +17,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findUserById(Long id) {
+    public User findUserById(int id) {
         return userRepo.findById(id).get();
     }
